prephr-backend/requirements.txt
Django==3.0.5
django-cors-headers==3.5.0
djongo==1.3.3
python-dotenv==0.14.0
sqlparse==0.2.4
pymongo==3.11.0
djangorestframework==3.11.0
numpy >= 1.16.1
opencv_python >= 3.4.8.29
dlib >= 19.16.0
social-auth-app-django == 4.0.0
social-auth-core >= 3.1.1
dnspython==2.0.0
